==> 06-interactive.Rmd <==
# Interactive component

For our interactive component, we have provided an exercise to test both your intuition and you understanding of our results. We have collected the top 100 days in arrests since 2011, and ranked them in order from most arrests to least. Can you determine the details of the number 1 date in arrests? Begin by selecting an answer to one of the white answer boxes below.

<script src="https://d3js.org/d3.v7.js"></script>

<script>
  
  // Setup and other initialization
  
  var w = 1500;
  var h = 700;
  
  var barchartMargin = {top: 50, right: 700, bottom: 50, left: 50};
  var barchartWidth = w - barchartMargin.left - barchartMargin.right;
  var barchartHeight = h - barchartMargin.top - barchartMargin.bottom;
  var dur = 500;
  
  var questionMargin = {top: 50, right: 50, bottom: 50, left: 750};
  var questionWidth = w - questionMargin.left - questionMargin.right;
  var questionHeight = h - questionMargin.top - questionMargin.bottom;

  // Initialize SVG and Background

  var svg = d3.select("body").select("div:nth-child(2)").select("div").select("div.page-wrapper")
    .append("svg")
    .attr("width", w)
    .attr("height", h);

  svg.append("rect")
      .attr("x", 0)
      .attr("y", 0)
      .attr("width", w)
      .attr("height", h)
      .attr("fill", "lightgray");
      
  // Initialize Scales and Axes in Global
  
  var xScale = d3.scaleLinear()
    .range([0, barchartWidth]);
  var yScale = d3.scaleBand()
    .range([0, barchartHeight])
    .paddingInner(.1);
  
  var xAxis = d3.axisBottom()
    .scale(xScale);
  var yAxis = d3.axisLeft()
    .scale(yScale);
    
  var chartTitle = svg.append("g")
    .attr("id", "chartTitles")
    .attr("transform", `translate (${barchartWidth / 2}, ${barchartMargin.top / 1.5}) scale(1.5)`)
    .append("text")
    .attr('text-anchor', 'middle')
    .text("Days with Highest Arrests - 2011-2021")
    
  var questionTitle = svg.append("g")
    .attr("id", "qTitles")
    .attr("transform", `translate (${barchartWidth + (questionWidth / 2)}, ${questionMargin.top / 1.5}) scale(1.5)`)
    .append("text")
    .attr('text-anchor', 'middle')
    .text("Questions")
  
  var xLabel = svg.append("g")
    .attr("id", "xLabs")
    .attr("transform", `translate (${barchartWidth / 2}, ${h - barchartMargin.bottom / 3})`)
    .append("text")
    .attr('text-anchor', 'middle')
    .text("Arrest Count")
  
  var yLabel = svg.append("g")
    .attr("id", "yLabs")
    .attr("transform", `translate (${barchartMargin.left / 2}, ${h / 2})`)
    .append("text")
    .attr('text-anchor', 'middle')
    .attr("transform", 'rotate(-90)')
    .text("Ranking")
  
  // Arrest Initial Questions
  
  var question = svg.append("g")
      .attr("id", "question")
      .append("rect")
      .attr("x", barchartWidth + barchartMargin.left + 50)
      .attr("y", barchartMargin.top + 20)
      .attr("width", questionWidth - 100)
      .attr("height", questionHeight / 5)
      .attr("fill","ivory");
						
	svg.select("g#question")
      .append("text")
      .attr("x", barchartWidth + barchartMargin.left + 60)
      .attr("y", barchartMargin.top + 50)
      .text("Which day of the week do you think sees the most arrests?");
  svg.select("g#question")
      .append("text")
      .attr("x", barchartWidth + barchartMargin.left + 60)
      .attr("y", barchartMargin.top + 110)
      .text("Fun Fact: Only one day in the top 100 does not fall on one of these weekdays!");
  svg.select("g#question")
      .append("text")
      .attr("x", barchartWidth + barchartMargin.left + 60)
      .attr("y", barchartMargin.top + 130)
      .text("It is Saturday, April 2nd, 2011, ranked 58th.");
  
  var q1answers = ["Tuesday","Wednesday","Thursday","Friday"]
    .map((value, key) => ({key, value}))
    
  var arrestQuestions = svg.append("g")
      .attr("id", "question_panels")
      .selectAll("rect")
      .data(q1answers, d => d.key)
      .enter()
      .append("rect")
      .attr("x", barchartWidth + barchartMargin.left + 50)
      .attr("y", (d, i) => barchartMargin.top + 200 + i * 100)
      .attr("width", questionWidth - 100)
      .attr("height", questionHeight / 10)
      .attr("fill","white")
      .attr("text",d => d.value)
      .on("click", function ()
					{
						question1update(d3.select(this).attr("text"))
						});
						
	svg.select("g#question_panels")
			.selectAll("text")
      .data(q1answers, d => d.key)
      .enter()
      .append("text")
      .attr("x", barchartWidth + barchartMargin.left + (questionWidth - 100)/2)
      .attr("y", (d, i) => barchartMargin.top + 200 + i * 100 + (questionHeight / 20))
      .text(d => d.value);  
  
  
  // Initialize Row Converter Function

  var rowConverter = function (d) {
    return {
      key: +d.Ranking,
      date: d.ARREST_DATE,
      weekday: d.Weekday,
      arrest_count: +d.Number_Arrests,
      year: d.ARREST_YEAR,
      month_num: +d.ARREST_MONTH,
      month_text: d.ARREST_MONTH_TEXT,
      precip: +d.Precipitation,
      temp: +d.Mean_Temperature,
    }
    
  };

  // Data Import Function

  d3.csv("https://raw.githubusercontent.com/Tylermarshall21/NYCArrests/main/interactive_testing/tophundred.csv", rowConverter)
  .then(function(arrestdata) {
    
    // Initialize Barchart Scale, Axes
    
    var xScale = d3.scaleLinear()
      .domain([0, 1750 /*d3.max(arrestdata, d => d.arrest_count)*/])
      .range([0, barchartWidth]);
      
    var yScale = d3.scaleBand()
      .domain(d3.range(1,arrestdata.length+1))
      .range([0, barchartHeight])
      .paddingInner(.1);
    
    var xAxis = d3.axisBottom()
      .scale(xScale);
      
    var yAxis = d3.axisLeft()
      .scale(yScale);
      
    // Create Bars
    
    var arrestBars = svg.append("g")
      .attr("id", "bars")
      .selectAll("rect")
      .data(arrestdata, d => d.key);
    
    arrestBars
      .enter()
      .append("rect")
      .attr("x", d => 0 + barchartMargin.left)
      .attr("y", (d, i) => yScale(d.key) + barchartMargin.top)
      .attr("width", d => xScale(d.arrest_count))
      .attr("height", yScale.bandwidth())
      .attr("fill","green")
      .text(d => `key: ${d.key}`);
    
    svg.append("g")
      .attr("class", "xAxis")
      .attr("transform", `translate (${barchartMargin.left}, ${h - barchartMargin.bottom})`)
      .call(xAxis);
      
    svg.append("g")
      .attr("class", "yAxis")
      .attr("transform", `translate (${barchartMargin.left}, ${barchartMargin.top})`)
      .call(yAxis);

    console.log("Nothing to see here!")    

  })
  .catch(function(error) {
    // error handling
    console.log("Something Wrong Here...")
  
  });

  

  // Update Function

  function update(arrestdata) {
    
    console.log(arrestdata.length)

    xScale.domain([0, 1750 /*d3.max(arrestdata, d => d.arrest_count)*/]);

    yScale.domain(d3.range(d3.min(arrestdata, d => d.key),d3.max(arrestdata, d => d.key)+1));

    var arrestBars = d3.select("g#bars")
				.selectAll("rect")
				.data(arrestdata, d => d.key)
				.text(d => `key: ${d.key}`);

    arrestBars
      .enter()
        .append("rect")
        .attr("x", d => 0 + barchartMargin.left)
        .attr("y", (d, i) => yScale(d.key) + barchartMargin.top)
        .attr("width", d => xScale(d.arrest_count))
        .attr("height", yScale.bandwidth())
        .text(d => `key: ${d.key}`)
      .merge(arrestBars)
        .transition()
        .duration(dur)
        .transition()
        .duration(dur)
        .ease(d3.easeLinear)
        .attr("x", d => 0 + barchartMargin.left)
        .attr("y", (d, i) => yScale(d.key) + barchartMargin.top)
        .attr("width", d => xScale(d.arrest_count))
        .attr("height", yScale.bandwidth())
        .text(d => `key: ${d.key}`);

    arrestBars.exit()
      .transition()
        .duration(dur)
        .attr("fill", "red")
        .attr("width", d => 0)
        .remove();

    svg.select(".xAxis")
      .transition()
        .duration(dur)
        .transition()
        .duration(dur)
        .ease(d3.easeLinear)
        .call(xAxis);

    svg.select(".yAxis")
      .transition()
        .duration(dur)
        .transition()
        .duration(dur)
        .ease(d3.easeLinear)
        .call(yAxis);

  }

  function question1update(weekday){
    console.log(weekday)
    
			var arrestBars = d3.select("g#bars")
			.selectAll("rect");

			var newarrestdata = arrestBars
				.data().filter(d => d.weekday == weekday);

				update(newarrestdata)
				if(newarrestdata.length == 0){
				  gameover(newarrestdata)
				}else{
				  createq2()
				}
		};
		
	function question2update(year){
    console.log(year)
    
			var arrestBars = d3.select("g#bars")
			.selectAll("rect");
			
			console.log(arrestBars.data())

			var newarrestdata = arrestBars
				.data().filter(d => d.year == year);
				
				update(newarrestdata)
				if(newarrestdata.length == 0){
				  gameover(newarrestdata)
				}else{
				createq3()
				}
		};
		
	function question3update(month_num){
    console.log(month_num)
    
			var arrestBars = d3.select("g#bars")
			.selectAll("rect");
			
			console.log(arrestBars.data())

			var newarrestdata = arrestBars
				.data().filter(d => (d.month_num <= month_num & d.month_num > month_num-3));
				
				update(newarrestdata)
				if(newarrestdata.length == 0){
				  gameover(newarrestdata)
				}else{
				createq4()
				}
		};
		
	function question4update(precip_thresh){
    console.log(precip_thresh)
    
			var arrestBars = d3.select("g#bars")
			.selectAll("rect");
			
			console.log(arrestBars.data())

			if(precip_thresh == 0.002){
			 var newarrestdata = arrestBars
				.data().filter(d => d.precip >= precip_thresh);
			}else{
			  var newarrestdata = arrestBars
				.data().filter(d => d.precip == precip_thresh);
			}
				
				update(newarrestdata)
				if(newarrestdata.length == 0){
				  gameover(newarrestdata)
				}else{
				createq5()
				}
		};
	
	function question5update(t_f){
    console.log(t_f)
    
			var arrestBars = d3.select("g#bars")
			.selectAll("rect");
			
			var best = arrestBars.data()[0].key
			var worst = arrestBars.data()[arrestBars.data().length - 1].key
			
			console.log(arrestBars.data())

			if(t_f == 0){
			 var newarrestdata = arrestBars
				.data().filter(d => d.key == best);
			}else{
			  var newarrestdata = arrestBars
				.data().filter(d => d.key == worst);
			}
				
				update(newarrestdata)
				gameover(newarrestdata)
		};
		
	function createq2(){
		  
		  // Arrest Second Round Questions
		  
		  svg.select("g#question").selectAll("text").remove();
		  svg.select("g#question").selectAll("rect").remove();

		  svg.select("g#question_panels").selectAll("rect").remove();
		  svg.select("g#question_panels").selectAll("text").remove();

  
  var question = svg.select("g#question")
      .append("rect")
      .attr("x", barchartWidth + barchartMargin.left + 50)
      .attr("y", barchartMargin.top + 20)
      .attr("width", questionWidth - 100)
      .attr("height", questionHeight / 5)
      .attr("fill","ivory");
						
	svg.select("g#question")
      .append("text")
      .attr("x", barchartWidth + barchartMargin.left + 60)
      .attr("y", barchartMargin.top + 50)
      .text("In which year do you think the highest arrest day is located?");
  svg.select("g#question")
      .append("text")
      .attr("x", barchartWidth + barchartMargin.left + 60)
      .attr("y", barchartMargin.top + 110)
      .text("Looking for other years? Believe it or not, none of them are represented in the top 100!");
  
  var q2answers = ["2011","2012","2013","2014"]
    .map((value, key) => ({key, value}))
    
  var arrestQuestions = svg.select("g#question_panels")
      .selectAll("rect")
      .data(q2answers, d => d.key)
      .enter()
      .append("rect")
      .attr("x", barchartWidth + barchartMargin.left + 50)
      .attr("y", (d, i) => barchartMargin.top + 200 + i * 100)
      .attr("width", questionWidth - 100)
      .attr("height", questionHeight / 10)
      .attr("fill","white")
      .attr("text",d => d.value)
      .on("click", function ()
					{
						question2update(d3.select(this).attr("text"))
						});
						
	svg.select("g#question_panels")
			.selectAll("text")
      .data(q2answers, d => d.key)
      .enter()
      .append("text")
      .attr("x", barchartWidth + barchartMargin.left + (questionWidth - 100)/2)
      .attr("y", (d, i) => barchartMargin.top + 200 + i * 100 + (questionHeight / 20))
      .text(d => d.value);
			
		};
		
	function createq3(){
		  
		  // Arrest Third Round Questions
		  
		  svg.select("g#question").selectAll("text").remove();
		  svg.select("g#question").selectAll("rect").remove();

		  svg.select("g#question_panels").selectAll("rect").remove();
		  svg.select("g#question_panels").selectAll("text").remove();

  
  var question = svg.select("g#question")
      .append("rect")
      .attr("x", barchartWidth + barchartMargin.left + 50)
      .attr("y", barchartMargin.top + 20)
      .attr("width", questionWidth - 100)
      .attr("height", questionHeight / 5)
      .attr("fill","ivory");
						
	svg.select("g#question")
      .append("text")
      .attr("x", barchartWidth + barchartMargin.left + 60)
      .attr("y", barchartMargin.top + 50)
      .text("In which months do your think the highest arrest days occurred?");
  svg.select("g#question")
      .append("text")
      .attr("x", barchartWidth + barchartMargin.left + 60)
      .attr("y", barchartMargin.top + 110)
      .text("Here's a hint, which month is it definitely not?");
  
  var q3answers = [{key:3, value:"January, February, March"},
                    {key:6, value:"April, May, June"},
                    {key:9, value:"July, August, September"},
                    {key:12, value:"October, November,December"}]
    
  var arrestQuestions = svg.select("g#question_panels")
      .selectAll("rect")
      .data(q3answers, d => d.key)
      .enter()
      .append("rect")
      .attr("x", barchartWidth + barchartMargin.left + 50)
      .attr("y", (d, i) => barchartMargin.top + 200 + i * 100)
      .attr("width", questionWidth - 100)
      .attr("height", questionHeight / 10)
      .attr("fill","white")
      .attr("text",d => d.value)
      .attr("key", d=> d.key)
      .on("click", function ()
					{
						question3update(d3.select(this).attr("key"))
						});
						
	svg.select("g#question_panels")
			.selectAll("text")
      .data(q3answers, d => d.key)
      .enter()
      .append("text")
      .attr("x", barchartWidth + barchartMargin.left + (questionWidth - 100)/2)
      .attr("y", (d, i) => barchartMargin.top + 200 + i * 100 + (questionHeight / 20))
      .text(d => d.value);
			
		};

  function createq4(){
		  
		  // Arrest Fourth Round Questions
		  
		  svg.select("g#question").selectAll("text").remove();
		  svg.select("g#question").selectAll("rect").remove();

		  svg.select("g#question_panels").selectAll("rect").remove();
		  svg.select("g#question_panels").selectAll("text").remove();

  
  var question = svg.select("g#question")
      .append("rect")
      .attr("x", barchartWidth + barchartMargin.left + 50)
      .attr("y", barchartMargin.top + 20)
      .attr("width", questionWidth - 100)
      .attr("height", questionHeight / 5)
      .attr("fill","ivory");
						
	svg.select("g#question")
      .append("text")
      .attr("x", barchartWidth + barchartMargin.left + 60)
      .attr("y", barchartMargin.top + 50)
      .text("On the days with the most arrests, what was the weather like?");
  svg.select("g#question")
      .append("text")
      .attr("x", barchartWidth + barchartMargin.left + 60)
      .attr("y", barchartMargin.top + 110)
      .text("Hint: Use your intuition!");
  
  var q4answers = [{key:0.00, value:"No precipitation"},
                    {key:0.001, value:"Trace amounts of precipitation"},
                    {key:0.002, value:"Precipitation greater than trace amounts"}]
    
  var arrestQuestions = svg.select("g#question_panels")
      .selectAll("rect")
      .data(q4answers, d => d.key)
      .enter()
      .append("rect")
      .attr("x", barchartWidth + barchartMargin.left + 50)
      .attr("y", (d, i) => barchartMargin.top + 200 + i * 100)
      .attr("width", questionWidth - 100)
      .attr("height", questionHeight / 10)
      .attr("fill","white")
      .attr("text",d => d.value)
      .attr("key", d=> d.key)
      .on("click", function ()
					{
						question4update(d3.select(this).attr("key"))
						});
						
	svg.select("g#question_panels")
			.selectAll("text")
      .data(q4answers, d => d.key)
      .enter()
      .append("text")
      .attr("x", barchartWidth + barchartMargin.left + (questionWidth - 100)/2)
      .attr("y", (d, i) => barchartMargin.top + 200 + i * 100 + (questionHeight / 20))
      .text(d => d.value);
			
		};
		
  function createq5(){
		  
		  // Arrest Fifth Round Questions
		  
		  svg.select("g#question").selectAll("text").remove();
		  svg.select("g#question").selectAll("rect").remove();

		  svg.select("g#question_panels").selectAll("rect").remove();
		  svg.select("g#question_panels").selectAll("text").remove();

  
  var question = svg.select("g#question")
      .append("rect")
      .attr("x", barchartWidth + barchartMargin.left + 50)
      .attr("y", barchartMargin.top + 20)
      .attr("width", questionWidth - 100)
      .attr("height", questionHeight / 5)
      .attr("fill","ivory");
						
	svg.select("g#question")
      .append("text")
      .attr("x", barchartWidth + barchartMargin.left + 60)
      .attr("y", barchartMargin.top + 40)
      .text("Final Question: Get this one right and we'll eliminate all but the top entry!");
	svg.select("g#question")
      .append("text")
      .attr("x", barchartWidth + barchartMargin.left + 60)
      .attr("y", barchartMargin.top + 60)
      .text("True or False: When crowds fill the city on Thanksgiving Day, daily arrests increase");
	svg.select("g#question")
      .append("text")
      .attr("x", barchartWidth + barchartMargin.left + 60)
      .attr("y", barchartMargin.top + 80)
      .text("significantly.");
  svg.select("g#question")
      .append("text")
      .attr("x", barchartWidth + barchartMargin.left + 60)
      .attr("y", barchartMargin.top + 130)
      .text("Hint: Think back to the analyses in the Results section.");
  
  var q4answers = [{key:1, value:"True"},
                    {key:0, value:"False"}]
    
  var arrestQuestions = svg.select("g#question_panels")
      .selectAll("rect")
      .data(q4answers, d => d.key)
      .enter()
      .append("rect")
      .attr("x", barchartWidth + barchartMargin.left + 50)
      .attr("y", (d, i) => barchartMargin.top + 200 + i * 100)
      .attr("width", questionWidth - 100)
      .attr("height", questionHeight / 10)
      .attr("fill","white")
      .attr("text",d => d.value)
      .attr("key", d=> d.key)
      .on("click", function ()
					{
						question5update(d3.select(this).attr("key"))
						});
						
	svg.select("g#question_panels")
			.selectAll("text")
      .data(q4answers, d => d.key)
      .enter()
      .append("text")
      .attr("x", barchartWidth + barchartMargin.left + (questionWidth - 100)/2)
      .attr("y", (d, i) => barchartMargin.top + 200 + i * 100 + (questionHeight / 20))
      .text(d => d.value);
			
		};
		
  function gameover(result_data){
		  
		  // Game is over - only 0 or 1 bar left
		  
		  svg.select("g#question").selectAll("text").remove();
		  svg.select("g#question").selectAll("rect").remove();

		  svg.select("g#question_panels").selectAll("rect").remove();
		  svg.select("g#question_panels").selectAll("text").remove();
		  
		  svg.select("g#qTitles").selectAll("text").remove();

  
  var question = svg.select("g#question")
      .append("rect")
      .attr("x", barchartWidth + barchartMargin.left + questionWidth / 2 - 175)
      .attr("y", barchartMargin.top + 30)
      .attr("width", questionWidth - 400)
      .attr("height", questionHeight / 6)
      .attr("fill","lightblue");
						
	svg.select("g#question")
      .append("text")
      .attr("x", barchartWidth + questionWidth/2 - 100)
      .attr("y", barchartMargin.top + 100)
      .style("font-size", "50px")
      .text("Game Over!");
  
  if(result_data.length == 1){
    
    var result_vals = result_data[0]
    
  var arrestQuestions = svg.select("g#question_panels")
      .append("rect")
      .attr("x", barchartWidth + barchartMargin.left + 50)
      .attr("y", (d, i) => barchartMargin.top + 200 + i * 100)
      .attr("width", questionWidth - 100)
      .attr("height", questionHeight / 10 - 10)
      .attr("fill","white")
						
	svg.select("g#question_panels")
      .append("text")
      .attr("x", barchartWidth + barchartMargin.left + 60)
      .attr("y", (d, i) => barchartMargin.top + (questionHeight / 20) + 200)
      .style("font-size", "30px")
      .text("You eliminated all but one date.");
      
  svg.select("g#question_panels")
      .append("rect")
      .attr("x", barchartWidth + barchartMargin.left + 50)
      .attr("y", (d, i) => barchartMargin.top + 220 + questionHeight/10)
      .attr("width", questionWidth - 100)
      .attr("height", questionHeight / 5)
      .attr("fill","white")
						
	svg.select("g#question_panels")
      .append("text")
      .attr("x", barchartWidth + barchartMargin.left + 60)
      .attr("y", (d, i) => barchartMargin.top + (questionHeight / 20) + 270)
      .text("Your selection had " + result_vals.arrest_count + " arrests!");
      
	svg.select("g#question_panels")
      .append("text")
      .attr("x", barchartWidth + barchartMargin.left + 60)
      .attr("y", (d, i) => barchartMargin.top + (questionHeight / 20) + 300)
      .text("Here are some other details about your selection:");
      
	svg.select("g#question_panels")
      .append("text")
      .attr("x", barchartWidth + barchartMargin.left + 60)
      .attr("y", (d, i) => barchartMargin.top + (questionHeight / 20) + 320)
      .text("Your selected date was: " + result_vals.date + ", ranked number " + result_vals.key);
      
	svg.select("g#question_panels")
      .append("text")
      .attr("x", barchartWidth + barchartMargin.left + 60)
      .attr("y", (d, i) => barchartMargin.top + (questionHeight / 20) + 340)
      .text("It was a " + result_vals.weekday);
  
  if(result_vals.precip == 0){
	svg.select("g#question_panels")
      .append("text")
      .attr("x", barchartWidth + barchartMargin.left + 60)
      .attr("y", (d, i) => barchartMargin.top + (questionHeight / 20) + 360)
      .text("There was no precipitation that day, and the average temperature was " + result_vals.temp + "°F");
  }else if(result_vals.precip == 0.001){
	svg.select("g#question_panels")
      .append("text")
      .attr("x", barchartWidth + barchartMargin.left + 60)
      .attr("y", (d, i) => barchartMargin.top + (questionHeight / 20) + 360)
      .text("There was trace precipitation that day, and the average temperature was " + result_vals.temp + "°F");
  }else{
	svg.select("g#question_panels")
      .append("text")
      .attr("x", barchartWidth + barchartMargin.left + 60)
      .attr("y", (d, i) => barchartMargin.top + (questionHeight / 20) + 360)
      .text("There was " + result_vals.precip + " inches of precipitation that day, and the average temperature was " + result_vals.temp + "°F");
  }
  // All possibilities eliminated...
		}else{
		  var arrestQuestions = svg.select("g#question_panels")
      .append("rect")
      .attr("x", barchartWidth + barchartMargin.left + 50)
      .attr("y", (d, i) => barchartMargin.top + 200 + i * 100)
      .attr("width", questionWidth - 100)
      .attr("height", questionHeight / 10 - 10)
      .attr("fill","white")
						
	svg.select("g#question_panels")
      .append("text")
      .attr("x", barchartWidth + barchartMargin.left + 60)
      .attr("y", (d, i) => barchartMargin.top + (questionHeight / 20) + 200)
      .style("font-size", "30px")
      .text("You eliminated all of the dates...");
		  
		}
		
		svg.select("g#question_panels")
      .append("rect")
      .attr("x", barchartWidth + barchartMargin.left + 50)
      .attr("y", h - 200)
      .attr("width", questionWidth - 100)
      .attr("height", questionHeight / 5 - 30)
      .attr("fill","yellow")
						
	svg.select("g#question_panels")
      .append("text")
      .attr("x", barchartWidth + barchartMargin.left + 60)
      .attr("y", h - 165)
      .style("font-size", "30px")
      .text("Not satisfied with your result?");
	svg.select("g#question_panels")
      .append("text")
      .attr("x", barchartWidth + barchartMargin.left + 60)
      .attr("y", h - 130)
      .style("font-size", "30px")
      .text("Refresh to try again!");
		};
    </script>
