==> 07-conclusion.Rmd <==
# Conclusion

In our project, we were able to explore various aspects of arrest and crime data from New York City from 2011-2021 and learn about how events and trends have shifted the landscape of law enforcement over that time. We were able to learn that arrests have definitely gone down during this time period, though crime data suggests not much has changed. That is not necessarily a good sign for safety in New York City, but is what the data is telling us. We were also able to see what the days of the week had the biggest arrest days, and discover that the top 50 days in terms of arrests were all either a Wednesday, a Thursday, or a Friday. Our other explorations allowed us to see the impact of laws changing in regards to marijuana arrests as well as the impact of drunk driving arrests after Uber and Lyft were introduced. Across many of our visuals we have seen the impact of the COVID-19 pandemic on arrest rates. At the close of our analysis, we have seen in great detail how events and trends, small and large, have played their part in the ever-evolving reality of law enforcement in the city, and we invite you to draw your own conclusions and insights as well.

## Analysis Limitations

Though we were able to learn a great deal while constructing this analysis, we also faced some limitations. The first limitation we had was that we had to limit our data to the last 10 years, due to the size of the dataset. This prevented us from being able to look at the impact of the recession on arrests, one of our initial areas of interest. Another limitation presented itself when we were comparing arrests with crime. The crime data we had access to was annual data, and lacked much of the geographic and demographic nuance found in our arrest data. As a result, we were only able to make simple year-to-year comparisons with the arrest data. We were unable to make month-by-month or day-by-day comparisons with the crime data to see if it was similar to some of the results we found with the arrest data, and we could not introduce many additional facets of interest into our analysis. One final nuisance was found in the form of age groups. Our arrest dataset split arrested individuals into 5 different age groups of varying sizes which made it hard to gain any insight from this variable. Though it was good to have any age data at all, it was unfortunate that the data was divided in such an uneven way.

## Lessons Learned

One lesson we learned from working with this data is that exploration results might not necessarily be easily expected or easily explained. When we created the Cleveland dot plot for the days with the most arrests, we were unable to identify why those days had the most arrests. We were hopeful that there might have been some events that could help explain what was going on that day, but we were unable to determine what happened to cause so many arrests. With the Thanksgiving Day parade heatmap, we thought arrests might be higher due to thefts around the parade, but this was not the case. In another analysis that was later cut from the final report, we tried to pair mundane events such as sports results to arrest spikes, but nothing materialized. In our analyses in general, we came to realize that knowledge of ongoing developments in law enforcement and legislation was critical to understanding shifts that might have otherwise been attributed to other causes. From this project, we learned that it is important to explore the data in many different ways, because the actual results might not match the expected results.

## Future Direction

In terms of future directions, we think it may be fruitful to look at the data from 2006-2010 to see how those 5 years compared to the last 10 years. This would allow us to look at the the impact of the recession on arrests while also giving us a larger picture of how New York City has been changing. We have noted a general downward trend in arrests since 2011, but we do not know if this continues further back in time. In addition, we could look to bring in some other datasets like we did with crime data to see if we could better explain some of our results. One set of information that comes to mind is a record of changes to police protocol during our window of exploration. Additionally, if we were able to find a dataset that contained all big events going on around the city, that data could potentially help us better explain our results in certain areas. Overall, we were able to learn a lot about arrests and crime in New York City throughout our exploration, but there is definitely still more to do with a dataset of this magnitude.
